Move form classes into a subnamespace
