+++ Form/Handler/NewThreadMessageFormHandler.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace FOS\MessageBundle\FormHandler;
+namespace FOS\MessageBundle\Form\Handler;
 
 use FOS\MessageBundle\FormModel\AbstractMessage;
 use FOS\MessageBundle\FormModel\NewThreadMessage;
@@ -10,9 +10,9 @@ class NewThreadMessageFormHandler extends AbstractMessageFormHandler
     /**
      * Composes a message from the form data
      *
-     * @param AbstractMessage $message
-     * @return MessageInterface the composed message ready to be sent
-     * @throws InvalidArgumentException if the message is not a NewThreadMessage
+     * @param  AbstractMessage          $message
+     * @return \FOS\MessageBundle\Model\MessageInterface         the composed message ready to be sent
+     * @throws \InvalidArgumentException if the message is not a NewThreadMessage
      */
     public function composeMessage(AbstractMessage $message)
     {
