+++ Form/Type/NewThreadMessageFormType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace FOS\MessageBundle\FormType;
+namespace FOS\MessageBundle\Form\Type;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
